Change how we name annotation targets for navigation properties
